# Conflicts:
#	src/GUI/Controllers/CreateMovieController.java
#	src/GUI/Controllers/MainController.java
#	src/GUI/Views/CreateMovieView.fxml

+++ src/GUI/Controllers/CategoriesController.java
@@ -2,6 +2,7 @@ package GUI.Controllers;
 
 import BE.Category;
 import GUI.Models.ModelsHandler;
+import GUI.Util.ConfirmOK;
 import GUI.Util.ExceptionHandler;
 import javafx.beans.InvalidationListener;
 import javafx.beans.Observable;
@@ -66,7 +67,8 @@ public class CategoriesController extends BaseController {
     private void handleRemoveCat(ActionEvent actionEvent) {
         Category catToDelete = tbvCat.getSelectionModel().getSelectedItem();
 
-        if (catToDelete != null) {
+        if (catToDelete != null && ConfirmOK.confirm("Remove category?",
+                        "Are you sure you want to remove " + catToDelete.getName() + "?")) {
             try {
                 getModelsHandler().getCategoryModel().deleteCategory(catToDelete);
                 getModelsHandler().getMovieModel().updateMovieCats(catToDelete);
